create auth api

+++ src/pre-bundle.js
@@ -6,6 +6,8 @@ import {
     uploadImageAndGetURL,
     addSubscriber,
 } from "../functions/models.mjs";
+import { signInWithCustomToken } from "firebase/auth";
+import { auth } from "../functions/database.mjs";
 
 // SUBSCRIBE FORM
 
@@ -27,6 +29,12 @@ async function handleSubscribe(e) {
 
 addEditFormHandlers();
 
+async function authenticate() {
+    const { token } = await fetch("/auth").then((res) => res.json());
+    !token && alert("Not authenticated!");
+    return await signInWithCustomToken(auth, token);
+}
+
 function addEditFormHandlers() {
     const changeImage = document.querySelector("#change-image"),
         uploadButton = document.querySelector("#upload-post"),
@@ -42,45 +50,55 @@ function addEditFormHandlers() {
 }
 
 async function handleChangeImage(e) {
-    const { post_id } = getFormData(),
-        file = e.target.files[0],
-        url = await uploadImageAndGetURL(post_id, file);
-    document.querySelector("#feature-image").value = url;
-    document.querySelector("#image-preview").src = url;
-    await handleAddDraft();
+    if (authenticate()) {
+        const { post_id } = getFormData(),
+            file = e.target.files[0],
+            url = await uploadImageAndGetURL(post_id, file);
+        document.querySelector("#feature-image").value = url;
+        document.querySelector("#image-preview").src = url;
+        await handleAddDraft();
+    }
     return;
 }
 
 async function handleAddPost() {
-    const post = getFormData();
-    await addPost(post).then((postAdded) => {
-        postAdded && alert("Post uploaded!");
-        window.location.href = "/admin";
-    });
+    if (authenticate()) {
+        const post = getFormData();
+        await addPost(post).then((postAdded) => {
+            if (postAdded) {
+                alert("Post uploaded!");
+                window.location.href = "/admin";
+            }
+        });
+    }
     return;
 }
 
 async function handleAddDraft() {
-    const post = getFormData();
-    await addDraft(post);
-    alert("Draft uploaded!");
+    if (authenticate()) {
+        const post = getFormData();
+        await addDraft(post);
+        alert("Draft uploaded!");
+    }
     return;
 }
 
 async function handleDeletePost() {
-    const { post_id } = getFormData(),
-        deletingPost = confirm(`Delete post with ID: ${post_id}`),
-        deletingDraft = confirm(`Delete draft with ID: ${post_id}`);
-    deletingPost && (await deletePost(post_id));
-    deletingDraft && (await deleteDraft(post_id));
-    if (deletingPost || deletingDraft) {
-        const message = `${capitalize(
-            [deletingPost && "post", deletingDraft && "draft"]
-                .filter(Boolean)
-                .join(" and ")
-        )} deleted!`;
-        alert(message);
-        window.location.href = "/admin";
+    if (authenticate()) {
+        const { post_id } = getFormData(),
+            deletingPost = confirm(`Delete post with ID: ${post_id}`),
+            deletingDraft = confirm(`Delete draft with ID: ${post_id}`);
+        deletingPost && (await deletePost(post_id));
+        deletingDraft && (await deleteDraft(post_id));
+        if (deletingPost || deletingDraft) {
+            const message = `${capitalize(
+                [deletingPost && "post", deletingDraft && "draft"]
+                    .filter(Boolean)
+                    .join(" and ")
+            )} deleted!`;
+            alert(message);
+            window.location.href = "/admin";
+        }
     }
     return;
 }
